Working on sharable playback state
Movie playback may be buggy :)

+++ apps/movies/movie-playback-handler.js
@@ -34,41 +34,39 @@ exports.startPlayback = function(response, movieUrl, movieFile, platform) {
 		ExecConfig = {  maxBuffer: 9000*1024, env: process.env.ffmpegPath };
 	}
 
-	if(fs.existsSync(outputPath) === true){
-		fs.unlinkSync(outputPath);
-	};   
-	    
-	GetMovieDurarion(response, movieUrl, movieFile, platform, function(data){
-		
-		response.json(data);  
-		
-		switch (platform) {
-			case "browser":
-				var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -preset ultraFast -acodec copy -sc_threshold 0 -movflags +frag_keyframe+empty_moov '+outputPath
-				break;
-			case "android":
-				var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -sc_threshold 0 -vb 250k -s 1280x720 -profile:v baseline -keyint_min 150 -deinterlace -c:a mp3 -movflags +frag_keyframe+empty_moov '+outputPath
-				break;
-			case "ios":
-				var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -preset ultraFast -acodec mp3 -ac 2 -ab 160k -sc_threshold 0 -movflags +frag_keyframe+empty_moov '+outputPath
-				break;
-			default:
-				console.log("Unknown platform: " + platform .red);
-				break;
-		}
-		
-		
-		var exec = require('child_process').exec
-			, child = exec(ffmpeg, ExecConfig, function(err, stdout, stderr) {
-				if (err) {
-					console.log('FFMPEG error: ',err) ;
-				} else{
-					console.log('Transcoding complete');
-				}
-			});
 
-		child.stdout.on('data', function(data) { console.log(data.toString()); });
-		child.stderr.on('data', function(data) { console.log(data.toString()); });			
+    GetMovieDurarion(response, movieUrl, movieFile, platform, function(data){
+        var movieDuration = data;
+        checkProgression(movieFile, function(progression){
+
+            if(progression !== 0 && progression !== undefined){
+                var movieProgression = progression;
+
+                if(fs.existsSync(outputPath) === false){
+                    // Assume file was transcoded.
+                    //TODO: Add check to check if transcoding was completed.
+                    startTranscoding(platform, movieUrl, outputPath, ExecConfig);
+                };
+
+            } else {
+                var movieProgression = 0;
+
+                // Cleanup just to be safe
+                if(fs.existsSync(outputPath) === true){
+                    fs.unlinkSync(outputPath);
+                };
+                startTranscoding(platform, movieUrl, outputPath, ExecConfig);
+            }
+
+            var movieInfo = {
+                'platform': platform,
+                'duration': movieDuration,
+                'progression': movieProgression
+            }
+
+            response.json(movieInfo);
+
+        });
 	});
 };
 
@@ -80,11 +78,8 @@ GetMovieDurarion = function(response, movieUrl, movieFile, platform, callback) {
         if(!err){
             if(probeData !== undefined || probeData.streams[0].duration !== 0 && probeData.streams[0].duration !== undefined && probeData.streams[0].duration !== "N/A" ){
                 console.log('Metadata found, continuing...');
-                var data = {
-                    'platform': platform,
-                    'duration':probeData.streams[0].duration
-                }
-               callback(data);
+                var data = probeData.streams[0].duration;
+                callback(data);
             } else {
                 console.log('Falling back to IMDB runtime information' .blue);
                 loadMetadataFromDatabase(movieFile, platform, function(data){
@@ -92,10 +87,7 @@ GetMovieDurarion = function(response, movieUrl, movieFile, platform, callback) {
                         callback(data);
                     } else{
                         console.log('Unknown movie duration, falling back to estimated duration.' .red);
-                        var data = {
-                            'platform': platform,
-                            'duration': 9000
-                        }
+                        var data = 9000;
                         callback(data);
                     }
                 });
@@ -103,10 +95,7 @@ GetMovieDurarion = function(response, movieUrl, movieFile, platform, callback) {
             }
         }else {
             console.log('Using fallback length due to error: ',err);
-            var data = {
-                'platform': platform,
-                'duration': 9000
-            }
+            var data = 9000;
             callback(data);
         }
 	});
@@ -118,12 +107,12 @@ loadMetadataFromDatabase = function(movieFile, platform, callback) {
 
 	db.query('SELECT * FROM movies WHERE original_name =? ', [ original_title ], {
 			local_name 		: String,
-			original_name  : String,
+			original_name   : String,
 			poster_path  	: String,
-			backdrop_path  : String,
+			backdrop_path   : String,
 			imdb_id  		: String,
-			rating  			: String,
-			certification  : String,
+			rating  		: String,
+			certification   : String,
 			genre  			: String,
 			runtime  		: String,
 			overview  		: String,
@@ -133,14 +122,58 @@ loadMetadataFromDatabase = function(movieFile, platform, callback) {
 			if (typeof rows !== 'undefined' && rows.length > 0){
 				var runtime = parseInt(rows[0].runtime) * 60;
 				console.log('Runtime found', rows[0].runtime);
-				var data = { 
-					'platform': platform,
-					'duration': runtime
-				}
+				var data = runtime;
 				callback(data);
 			} else {
 				callback(null);
 			}
 		}
 	);
+}
+
+checkProgression = function(movieFile, callback) {
+    db.query('SELECT * FROM progressionmarker WHERE movietitle =? ', [ movieFile ], {
+            movietitle 		: String,
+            progression     : String
+        },
+        function(rows) {
+            if (typeof rows !== 'undefined' && rows.length > 0){
+                var progression = rows[0].progression;
+                callback(progression);
+            } else {
+                callback(0);
+            }
+        }
+    );
+}
+
+
+startTranscoding = function(platform, movieUrl, outputPath, ExecConfig){
+    switch (platform) {
+        case "browser":
+            var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -preset ultraFast -acodec copy -sc_threshold 0 -movflags +frag_keyframe+empty_moov '+outputPath
+            break;
+        case "android":
+            var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -sc_threshold 0 -vb 250k -s 1280x720 -profile:v baseline -keyint_min 150 -deinterlace -c:a mp3 -movflags +frag_keyframe+empty_moov '+outputPath
+            break;
+        case "ios":
+            var ffmpeg = 'ffmpeg -i "'+movieUrl+'" -g 52 -threads 0 -vcodec libx264 -coder 0 -flags -loop -pix_fmt yuv420p -crf 22 -subq 0 -preset ultraFast -acodec mp3 -ac 2 -ab 160k -sc_threshold 0 -movflags +frag_keyframe+empty_moov '+outputPath
+            break;
+        default:
+            console.log("Unknown platform: " + platform .red);
+            break;
+    }
+
+
+    var exec = require('child_process').exec
+    , child = exec(ffmpeg, ExecConfig, function(err, stdout, stderr) {
+        if (err) {
+            console.log('FFMPEG error: ',err) ;
+        } else{
+            console.log('Transcoding complete');
+        }
+    });
+
+    child.stdout.on('data', function(data) { console.log(data.toString()); });
+    child.stderr.on('data', function(data) { console.log(data.toString()); });
 }
